ncp: fix handling of NDS List requests and replies.
In requests:

There appear to be 2 bytes of unknown data (typically 0) after the
2-byte Request Flags field (are they just 2 bytes of additional flags?).
Skip past them before dissecting the iterator.

If there are no bytes remaining in the packet after the parent ID, stop
dissecting; some packets seem to stop there.  For those requests, assume
that the response will contain :

	entry ID;
	entry flags;
	subordinate count;
	modification time;
	base class;
	relative distinguished name;

although the last of those might be something else (it appears to be of
the form "CN={name}").

In replies:

For each returned entry, if the requested field flags in the request had
the DSI_OUTPUT_FIELDS bit set, fetch the returned field flags and use
that to determine what fields are present; otherwise, use the requested
field flags.