STUN: add an option to auto-detect STUN protocol version
Commit 3a5d0569d7 added support for different STUN protocol versions;
a global preference allows the user to select the desired flavour.

Unfortunately, it is pretty common to have different flavours in the same
capture file, or even in the same packet (with STUN messages encapsulated
in a TURN tunnel), so a global preference applied to the entire file might
not provide enough flexibility.

Add a basic auto-detect algorithm to identify the STUN version specific to
each STUN message (the users will still have the option to force a global
version)