SMPP: Support GSM 7 bit alphabet (3GPP TS 23.038)
Support the GSM 7 bit alphabet. This should be unpacked, according
to the spec and references online. Similarly, the IA5/ASCII support
should also be unpacked. GSM 7 Packed is supported as an option for
encoding 0, as reports of a small number of servers behaving that way
exist. Provide the byte representation for when encoding is performed
(but hidden), so that filters doesn't change depending on the preference
and encoding, and are backwards compatible with pre 3.2.0. Fixes #17015