packet-tls-utils:Move ssl_sha_init() and ssl_md5_init() out of a loop
Instead of calling ssl_sha_init() and ssl_sha_cleanup() in
a loop, call the newly-created ssl_sha_reset(), to avoid
creating and releasing contexts.
Also, do the same for ssl_md5_init() and ssl_md5_cleanup().

Besides the performance improvement,
this should also fix Coverity 1450156 and 1450693,
which detect the multiple calls to ssl_sha_cleanup()
and ssl_md5_cleanup() as a double-free.