Qt: column list model: remove unused ListElement component
Selecting Edit / Preferences on my asan+ubsan build brings up the
following warning from ubsan:

/media/sf_wireshark.git/ui/qt/models/column_list_model.cpp:273:9:
runtime error: load of value 25, which is not a valid value for type 'bool'

The problem is in ColumnListModel::populate(), where a ListElement is
populated from fmt_data. The ListElement's "changed" component is not
initialized, though.

It looks like "changed" is not used anywhere. This patch removes it from
the ListElement struct.