wlan: fix wrong reported length for reassembled payloads
There is no need to remove the FCS at the end of the reassembled payload since FCS bytes have already been removed from the fragments. Also, the way it was removed in the code was wrong: it set the reported length of the reassembled payload to the reported length of the last fragment only.

Revert "ieee802.11: Make sure we do not include the FCS in the next_tvb."

This reverts commit 094297ec77a28a38310d9449cd4f700163434022.