Work around more NeXTStEP-induced pain.
GNU libtool has a libtool program and a libtoolize program.

The development tools for NeXTStEP, apparently, had a libtool program as
well, and the current version of the development tools for the current
version of NeXTStEP, generally known as "Xcode for macOS", still have
that program.

This means that we do some renaming after installing GNU libtool, so
that its "libtool" becomes "glibtool" and its "libtoolize" becomes
"glibtoolize".

That meant we had to compensate for that when running autoreconf when
building and installing minizip.

It turns out we have to do that when running autogen.sh when building
and installing GLib as well.


(cherry picked from commit 2ecb6061422ca933db2422e9ecb03bb0544a6492)