IPv6: SRH setup correct DA for pseudo header
When a Segment Routing Header is present in the IPv6 packet provisions
have to be made to setup the right destination address for the pseudo
header used in checksum calculations. When segments are left in the header
the first address in the list has to replace the destination address.

Closes #17097


(cherry picked from commit 7052994a19bd3361ea9181c6ef99a82b398033ed)