Qt: Debounce Preferences -> Advanced menu's "Search" to improve UX
As running pd_ui_->advancedView->expandAll() takes a noticeable amount
of time and so would introduce significant lag while typing a string
into the Search box, we instead debounce the call to
updateSearchLineEdit(), so that it doesn't run until a set amount of
time has elapsed with no updates to the Search field.

If the user types something before the timer elapses, the timer restarts
the countdown.

Fixes #17107