TFTP: Parse "windowsize" and "msftwindow" options.
Add parsing of RFC 7440 "TFTP Windowsize Option" and the Microsoft-proprietary
"msftwindow" option seen in Windows Deployment Services (WDS) PXE.  I have
found no spec or documentation for the latter, so it is reverse-engineered from
various packet captures of WDS boot.

In the RRQ, WRQ, and OACK, parse the "windowsize" and "msftwindow" options.  I
haven't seen "msftwindow" in an WRQ, only in an RRQ and corresponding OACK, but
I have implemented it symmetrically in the WRQ in this change.

In the ACK, parse the extra byte added when "msftwindow" is active (containing
the dynamic window size) and add it to the protocol tree.  Include a "chat"
expert info for changes in window size, since changes in packet queue depth
might indirectly trigger defects in a PXE implementation.