dumpcap: Improve the secondary error for promiscuous mode failure.
On Windows, some devices don't let promiscuous mode be enabled, and
return an error rather than silently ignoring the request to use
promiscuous mode (as UN*X devices tend to do).  Check for the error
message from that error, and suggest that the user turn off promiscuous
mode on that device.