TFTP: Link each DATA/ACK to its read/write request.
Let's keep track of the frame number which contained the initial read or write
request for the conversation.  This allows us to easily add a link from each
DATA or ACK packet back to the read or write request which started the
transfer.

While we are here, add a missing initialization of tftp_info->windowsize and
sort the initializers into the same order as the fields.